Moved sources to src directory, examples to examples directory, added Ferno\Loco namespace and fixed composer.json autoloading
