+++ examples/bnf.php
