+++ examples/regEx.php
