+++ examples/simpleComment.php
