+++ examples/wirth.php
